# backslash/swagger.yaml
openapi: '3.0.0'
info:
  version: '1.0.0'
  title: Home Exercise - Backslash
servers:
  - url: http://localhost:3000/
paths:
  /routes/sinks:
    get:
      summary: Get list of GraphPath object contains all the routes that ends in Sink
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphPathArray'
      tags:
        - End In Sink
  /routes/sinks/{sinkKind}:
    get:
      summary: Get list of GraphPath object contains all the routes that ends in Sink by filter kind
      parameters:
        - name: sinkKind
          in: path
          description: Kind of the Sink (sqs/rds)
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphPathArray'
      tags:
        - End In Sink
  /routes/vulnerabilities:
    get:
      summary: Get list of GraphPath object contains all the routes that have vulnerabilities
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphPathArray'
      tags:
        - Contains Vulnerability
  /routes/publicServiceToSink:
    get:
      summary: Get list of GraphPath object contains all the routes  that start in a public service and ends in Sink
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphPathArray'
      tags:
        -  Start In Public Service End In Sink Kind
  /routes/publicServiceToSink/{isPublic}/{sinkKind}:
    get:
      summary: Get list of GraphPath object contains all the routes that ends in Sink by filter kind
      parameters:
        - name: isPublic
          in: path
          description: public exposed
          required: true
          schema:
            type: boolean
        - name: sinkKind
          in: path
          description: Kind of the Sink (sqs/rds)
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphPathArray'
      tags:
        - Start In Public Service End In Sink Kind
components:
  schemas:
    GraphPathArray:
      type: array
      items:
        $ref: '#/components/schemas/GraphPath'
    GraphPath:
      type: object
      properties:
        source:
          type: string
          description: Source node of the path
        dest:
          type: string
          description: Destination node of the path
        segments:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    Link:
      type: object
      properties:
        source:
          $ref: '#/components/schemas/NodeI'
          description: Source node of the link
        dest:
          $ref: '#/components/schemas/NodeI'
          description: Destination node of the link
    NodeI:
      type: object
      properties:
        name:
          type: string
          description: Name of the node
        kind:
          type: string
          description: Kind of the node
        language:
          type: string
          description: Language of the node
        path:
          type: string
          description: Path of the node
        publicExposed:
          type: boolean
          description: Whether the node is publicly exposed
        vulnerabilities:
          type: array
          items:
            $ref: '#/components/schemas/VulnerabilityI'
          description: Vulnerabilities of the node
        metadata:
          $ref: '#/components/schemas/MetadataI'
          description: Metadata of the node
    VulnerabilityI:
      type: object
      properties:
        file:
          type: string
          description: File with vulnerability
        severity:
          type: string
          enum: [low, medium, high]
          description: Severity of the vulnerability
        message:
          type: string
          description: Description of the vulnerability
        metadata:
          $ref: '#/components/schemas/MetadataI'
          description: Metadata of the vulnerability
    MetadataI:
      type: object
      properties:
        cwe:
          type: string
          description: CWE of the metadata
        cloud:
          type: string
          description: Cloud of the metadata
        engine:
          type: string
          description: Engine of the metadata
        version:
          type: string
          description: Version of the metadata